+++ src/routes/graphql/schemas.ts
@@ -1,4 +1,6 @@
 import { Type } from '@fastify/type-provider-typebox';
+import { GraphQLSchema } from 'graphql';
+import { mutationType, queryType } from './schema.js';
 
 export const gqlResponseSchema = Type.Partial(
   Type.Object({
@@ -18,3 +20,7 @@ export const createGqlResponseSchema = {
     },
   ),
 };
+export const gqlSchema = new GraphQLSchema({
+  query: queryType,
+  mutation: mutationType,
+});
